UseBackgroundThreadsForIO should be true
This setting just means that having an open connection will not
keep the process alive. If the host otherwise shuts down, we
don't want just the connection to keep it alive. This setting only
impacts the desktop framework, and is likely to become obsolete
in the next version of the the RabbitMq client.