Use Push API to consume
This changes from a pull api to the prefellered push API
of RabbitMQ. This can substanctially inprove throughput.
Unfortunately, the orleans API is not really set up for
consuming a push based API with its own built-in
backpressure system, so we end up with both the cache
and a second in memory queue.

Ideally we would instead have a custom version of
SimpleQueueCache, without a max size, and no
backpressure. This would ensure that messages
spend as little time in the extra in-memory queue as
possible. Howver I'm not sure the extra maintenance
burden would be worth it.