Filter batches produced on wrong queue
Filter out any decoded batches whose streams that do not belong
to the queue. This can happen with custom data adapters, and unusual
typologies. If it does we don't want to try to deliver those because
those messages should also appear in the correct queue, and thus would
get needlessly double delivered.